requirements.txt
ultralytics==8.0.199
opencv-python
numpy
Pillow
requests
python-dotenv
replicate
gradio
